+++ tests/spl/token-extensions/programs/token-extensions/src/lib.rs
@@ -42,4 +42,12 @@ pub mod token_extensions {
     ) -> Result<()> {
         instructions::update_group_pointer_handler(ctx, new_group_address)
     }
+
+    pub fn enable_cpi_guard(ctx: Context<EnableCpiGuard>) -> Result<()> {
+        instructions::enable_cpi_guard_handler(ctx)
+    }
+
+    pub fn disable_cpi_guard(ctx: Context<DisableCpiGuard>) -> Result<()> {
+        instructions::disable_cpi_guard_handler(ctx)
+    }
 }